requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
pypdf
requests
python-multipart
streamlit